init commit
**FOLDERS** is a simple component that creates folders in the local directory
relative to the project folder. Since a number of my components have default
locations for externally synced tables, extensions etc., I use this to  define
and guarantee a consistent local file structure

To modify, customize this component and add any number of folder paramaters to
the Folders page.

Any folder that is defined here will be checked for on pulse, and created if it
doesn't already exist.

If onCreate is selected, when dropping this into a new project, it will
automatically create the desire folders.

If onStart is selected, it will check whenever the TDsession that it is in when
the project is launched.

this component is a part of HIVE, a collection of tools by @drmbt

[instagram.com/drmbt](instagram.com/drmbt)
[patreon.com/drmbt](patreon.com/drmbt)
[github.com/drmbt](github.com/drmbt)